treewide: format size_t using PRIu64
Currently we use %ld in some places to format size_t variables. This
may overflow and print negative numbers in some cases which can result
in confusing output for the end user.

Change the formatting strings to use %PRIu64 for size_t everywhere to
avoid that.  We've been using PRIu64 since 53614b1351 ("commit-graph:
verify parent list", 2018-06-27), and haven't heard any complaints, so
this macro should be safe to use everywhere.

Some places converted here might not actually realistically overflow.
However it's easier to convert everything at once instead of checking
whether or not ever callsite can or can't overflow.  More importantly,
this brings us one step closer to turning on '-Wformat-signedness',
which will avoid similar problems once and for all in the future.

+++ split-index.c
@@ -108,8 +108,8 @@ static void mark_entry_for_delete(size_t pos, void *data)
 {
 	struct index_state *istate = data;
 	if (pos >= istate->cache_nr)
-		die("position for delete %d exceeds base index size %u",
-		    (int)pos, istate->cache_nr);
+		die("position for delete %"PRIu64" exceeds base index size %u",
+		    pos, istate->cache_nr);
 	istate->cache[pos]->ce_flags |= CE_REMOVE;
 	istate->split_index->nr_deletions++;
 }
@@ -121,19 +121,19 @@ static void replace_entry(size_t pos, void *data)
 	struct cache_entry *dst, *src;
 
 	if (pos >= istate->cache_nr)
-		die("position for replacement %d exceeds base index size %u",
-		    (int)pos, istate->cache_nr);
+		die("position for replacement %"PRIu64" exceeds base index size %u",
+		    pos, istate->cache_nr);
 	if (si->nr_replacements >= si->saved_cache_nr)
 		die("too many replacements (%u vs %u)",
 		    si->nr_replacements, si->saved_cache_nr);
 	dst = istate->cache[pos];
 	if (dst->ce_flags & CE_REMOVE)
-		die("entry %d is marked as both replaced and deleted",
-		    (int)pos);
+		die("entry %"PRIu64" is marked as both replaced and deleted",
+		    pos);
 	src = si->saved_cache[si->nr_replacements];
 	if (ce_namelen(src))
-		die("corrupt link extension, entry %d should have "
-		    "zero length name", (int)pos);
+		die("corrupt link extension, entry %"PRIu64" should have "
+		    "zero length name", pos);
 	src->index = pos + 1;
 	src->ce_flags |= CE_UPDATE_IN_BASE;
 	src->ce_namelen = dst->ce_namelen;
